Change input conversion to float for accurate arithmetic in calculator.py

+++ calculator.py
@@ -8,6 +8,6 @@
 
 #or
 
-x= int(input("what's x? "))
-y= int(input("what's y? "))
+x=float(input("what's x? "))
+y= float(input("what's y? "))
 print(x+y)
